+++ ec/modules/helpers.py
@@ -48,25 +48,29 @@ def load_module(module_path):
   
   return imported
 
-KWARG_PATTERN = re.compile('.+(?<!\\\\)=')
+KWARG_VALIDATOR = re.compile('.+(?<!\\\\)=')
 def getDigestableArgs(Argv):
   """Splits the given Argv into *Args and **KwArgs.
   """
   first_kwarg_pos = 0
   
   for arg in Argv:
-    if KWARG_PATTERN.search(arg):
+    if KWARG_VALIDATOR.search(arg):
       break
       
     else:
       first_kwarg_pos += 1
   
   for arg in Argv[first_kwarg_pos:]: # ensure that the kwargs are valid
-    if not KWARG_PATTERN.search(arg):
+    if not KWARG_VALIDATOR.search(arg):
       raise HandledException('Could not parse the arg "%s".' % arg)
       
   return Argv[:first_kwarg_pos], list2dict(Argv[first_kwarg_pos:])
 
+NAME_VALIDATOR = re.compile('[\\w_]+')
+def validateName(name):
+  assert(NAME_VALIDATOR.search(name))
+  
 # mode helpers
 def getMemberHelp(Target):
   if isinstance(Target, Group):
@@ -138,7 +142,7 @@ def listMemberHelps(TargetGroup):
   return Ret
   
 def getGroupHelp(_Group):
-  return '\n\n'.join([('%s  %s' % (name, desc))[:60] for name, desc in listMemberHelps(_Group)])
+  return '\n\n'.join([('%s:  %s' % (name, desc))[:60] for name, desc in listMemberHelps(_Group)]) # ToDo: Better formatting of the help instead of chopping the text.
 
 def getAutoDesc(ArgConfig):
   return '%s (%s)' % (ArgConfig['name'], ArgConfig['default']) if 'default' in ArgConfig else ArgConfig['name']